move seq rdd test files
