+++ src/test/scala/org/hammerlab/magic/rdd/serde/SequenceFileRDDTest.scala
@@ -1,8 +1,9 @@
-package org.hammerlab.magic.test.rdd
+package org.hammerlab.magic.rdd.serde
 
 import org.apache.hadoop.io.compress.CompressionCodec
 import org.apache.spark.rdd.RDD
 import org.hammerlab.magic.rdd.serde.SequenceFileSerializableRDD._
+import org.hammerlab.magic.test.rdd.VerifyRDDSerde
 
 import scala.reflect.ClassTag
